==> 10_plots.Rmd <==
---
title: "10_plots"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(pacman)
p_load(tidyverse, scales, ggwordcloud, ggunchained)
```

# LOAD
```{r load novelty-transience-resonance}
## original data
full <- read_csv("data/csv/full_ref.csv") %>%
  select(-text_c, -speaker_index, -bod_index, -hhmm)

## lda_output + novelty, resonance
ntr <- list.files(path = "data/ntr", pattern = "_from_matrix", full.names = T)

read_csv_and_assign <- function(path) {
  subset <- read_csv(path) %>%
    mutate(group = str_extract(path, "\\d"))
}

comparison <- map_df(ntr, read_csv_and_assign) %>%
  rename(rowname = doc_id) %>%
  arrange(rowname) %>%
  left_join(full, by = "rowname")
```


# PLOTS
```{r metric plots}
# NOVELTY VS RESONANCE
ggplot(comparison, aes(z_novelty, z_resonance)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000),
                         labels = trans_format("log10", 
                                               math_format(expr = 10^.x, format = force))) +
    geom_smooth(method = "lm", colour = "black", alpha = 0, size = 0.5) +
    labs(x = "Novelty (z-scaled)\n", y = "Resonance (z-scaled)", 
         title = "Resonance vs. Novelty in the short-term",
         subtitle = "with a regression line, w = 27",
         caption = "Data source: speeches in Czech Chamber of Deputies (2010-2019)",
         fill = "speech\ncount") +
  scale_x_continuous(breaks = seq(-4, 8, 2)) +
  scale_y_continuous(breaks = seq(-6, 6, 2)) +
  theme_janco_point() +
  theme(legend.direction = "vertical", legend.position = "right",
        plot.caption = element_text(hjust = 0.5)
        )


# NOVELTY VS TRANSIENCE
ggplot(comparison, aes(z_novelty, z_transience)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    geom_abline(colour = "black", linetype = "dashed") +
    labs(x = "Novelty (z-scaled)\n", y = "Transience (z-scaled)", 
         title = "Novelty vs. Transience in the short-term",
         subtitle = "with an identity line (x = y), w = 27",
         caption = "Data source: speeches in Czech Chamber of Deputies (2010-2019)",
         fill = "speech\ncount") +
  scale_x_continuous(breaks = seq(-4, 8, 2)) +
  scale_y_continuous(breaks = seq(-6, 8, 2)) +
  theme_janco_point() +
  theme(legend.direction = "vertical", legend.position = "right",
        plot.caption = element_text(hjust = 0.5)
        )
```


BY PARTY
```{r party plots}
pc <- comparison %>%
  mutate(zkratka = ifelse(is.na(zkratka), "Guest Speakers", zkratka)) %>%
  mutate(party = case_when(zkratka == "ANO2011" ~ "Babiš et al.",
                           zkratka == "CSSD" ~ "SocDem",
                           zkratka == "KDU-ČSL" ~ "ChristDem",
                           zkratka == "KSCM" ~ "Commies",
                           zkratka == "ODS" ~ "Conservatives",
                           zkratka == "TOP09" ~ "Big-City-Liberals",
                           zkratka == "STAN" ~ "Village-Liberals",
                           zkratka == "Piráti" ~ "Pirates",
                           zkratka == "Usvit" | zkratka == "SPD" ~ "Nazis",
                           zkratka == "VV" ~ "Catch-all"))

# by party nov-res line
pc %>%
  #filter(group == 8) %>%
  ggplot(aes(novelty, resonance, color = party)) +
  geom_smooth(alpha = 0) +
  theme_janco_point()

# split-violin plot comparing bellow- and above- average novely to transience across parties
pc %>%
  mutate(hi_novelty = ifelse(z_novelty >= 0, 1, 0)) %>%
  ggplot(aes(party, transience, fill = factor(hi_novelty))) +
  geom_split_violin() +
  theme_janco_bar()

pc %>%
  filter(group == 4) %>%
  ggplot(aes(z_novelty, z_resonance)) +
  geom_bin2d(bins = 100) +
  scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
  geom_smooth(method = "lm", colour = "black", size = 0.5) +
  facet_wrap(~zkratka) +
  theme_janco_point() +
    labs(x = "Novelty (z-scaled)\n", y = "Resonance (z-scaled)", 
       title = "Resonance vs. Novelty by party: 2013/14",
       subtitle = "w = 27",
       caption = "Data source: speeches in Czech Chamber of Deputies (2013-2014)",
       fill = "speech count \n")
```

ANO vs CSSD
```{r populists plots}
ano <- pc %>%
  filter(zkratka == "ANO2011" | zkratka == "CSSD") %>%
  filter(group != 1 & group != 2 & group != 3)

ano %>%
  ggplot() +
  geom_jitter(aes(factor(group), z_novelty, color = zkratka), alpha = 0.01, position = position_jitterdodge()) +
  geom_split_violin(aes(factor(group), z_novelty, fill = zkratka, alpha = 0.8)) +
  scale_x_discrete(labels = c("2013/14", "2014/15", "2015/16", "2017/18", "2018/19")) +
  scale_color_manual(values = party_colors, aesthetics = c("fill", "color")) +
  theme_janco_bar() +
  guides(alpha = F) +
  labs(title = "Novelty over time",
       subtitle = "Comparing Andrej Babis' newcomer party and the Social Democrats, w = 27",
       y = "novelty (z-scaled)",
       x = "timeframe",
       color = "party",
       fill = "party")

ano %>%
  ggplot() +
  geom_jitter(aes(factor(group), z_resonance, color = zkratka), alpha = 0.01, position = position_jitterdodge()) +
  geom_split_violin(aes(factor(group), z_resonance, fill = zkratka, alpha = 0.8)) +
  scale_x_discrete(labels = c("2013/14", "2014/15", "2015/16", "2017/18", "2018/19")) +
  scale_color_manual(values = party_colors, aesthetics = c("fill", "color")) +
  theme_janco_bar() +
  guides(alpha = F) +
  labs(title = "Resonance over time",
       subtitle = "Comparing Andrej Babis' newcomer party and the Social Democrats, w = 27",
       y = "resonance (z-scaled)",
       x = "timeframe",
       color = "party",
       fill = "party")
```
